Move to proc-macro
