+++ module-system/sov-modules-macros/src/rpc/mod.rs
@@ -1,7 +1,5 @@
 mod expose_rpc;
-mod expose_rpc_2;
 mod rpc_gen;
 
-pub(crate) use expose_rpc::expose_rpc;
-pub(crate) use expose_rpc_2::ExposeRpcMacro;
+pub(crate) use expose_rpc::ExposeRpcMacro;
 pub(crate) use rpc_gen::rpc_gen;
